Final Commit
Ready for review. Added the selector at the bottom of the form for the registration category. Fixed spelling errors in the ids of some of the forms. Changed the address form an input box to a textarea to match the screenshot. Added in comments. Added in styles to make the input boxes change to yellow when selected. Added in validation styles for invalid and valid fields. Added in comments for the CSS code.